backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.18
PyMuPDF==1.23.8
python-multipart==0.0.6
